Isolate audio waveform extraction from engine facade
The waveform operation can live behind its own module while preserving mcp_video.engine.audio_waveform for client, server, CLI, and external callers. The move narrows engine.py and keeps the existing synthetic fallback behavior for FFmpeg astats failures.

Constraint: Preserve public import compatibility and existing CLI fallback behavior while continuing the approved engine decomposition plan.

Rejected: Treat nonzero astats return codes as hard failures | current fixtures rely on synthetic fallback when ametadata is unavailable.

Confidence: high

Scope-risk: narrow

Directive: Keep waveform synthetic fallback unless callers gain a separate strict-analysis mode.

Tested: ruff check mcp_video/engine.py mcp_video/engine_audio_waveform.py; ruff format --check mcp_video/engine.py mcp_video/engine_audio_waveform.py; python3 -m pytest tests/test_engine_advanced.py -k 'waveform' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'audio_waveform' -q --tb=short; python3 -m pytest tests/test_red_team.py -k 'audio_waveform' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
